readability updates
